Add fixes for RDS Postgres and some hasPermission bugs
RDS `postgres` user is more limited than a traditional `postgres` user
and needs some tricks.

The user creating a database the creator must be a member of the role
when setting another owner or setting default privileges. To get around
this we temporarily add the current user to that role.

Also found `has_database_privilege` returns `true` if any privileges are
set, not all. Assumed the other `has` functions work the same and
amended our functions to instead loop through desired permissions
instead.

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>