Refactor user creation and update logic in PostgreSQL manager
This adds additional options for a role/user which are applied when
creating or now updating a user. Also adds a `getUser` method to read
back a role's properties to make this a bit more idempotent and to
improve test assertions.

Updated `createUser` so that it specifies the LOGIN flag either when a
password is set or the `Options.Login` is explicitly set (this is needed
when configuring another authentication method, like IAM).

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>